move selectedFeatures to selected.js as .blockFeatures
draw-map.js : draw() :
move selectedFeatures_clear() to selected.js
use selectedService.blocksFeatures for selectedFeatures.

mapview.js : move .selectedFeatures to selected.js and alias from that.

selected.js :
move draw-map.js : draw() : selectedFeatures to here as .blocksFeatures.
add updateSelectedFeatures(), selectedFeatures_add().

factor from draw-map.js : selectedFeaturesSendArray() to form selectedBlocksFeaturesToArray();
use Object.keys in place of d3.keys on selectedFeatures.
