+++ frontend/app/services/data/selected.js
@@ -20,6 +20,26 @@ export default Service.extend(Evented, {
 
   // ---------------------------------------------------------------------------
 
+  /** Selected features, in a single array
+   */
+
+  /** array of {Chromosome, Feature, Position, feature}.
+   * selectedBlocksFeaturesToArray() returns this form.
+   */
+  selectedFeatures : [],
+
+  /** Selected features, grouped in per-block arrays.
+   * [blockId] -> [feature, ...]
+   * This is operated on by :
+   *  selectedFeatures_add(), selectedFeatures_clear(),
+   *  selectedBlocksFeaturesToArray(),
+   *  removeUnviewedBlockFeaturesFromSelected()
+   * The result of selectedBlocksFeaturesToArray() can be written to .selectedFeatures with updateSelectedFeatures.
+   */
+  blocksFeatures : {},
+
+  // ---------------------------------------------------------------------------
+
   selectedElements : A(),
 
   // ---------------------------------------------------------------------------
@@ -42,6 +62,9 @@ export default Service.extend(Evented, {
    * After feature search, labelledFeatures will be intersected with the feature search result.
    */
   labelledFeatures : A(),
+
+  // ---------------------------------------------------------------------------
+  /** Operations on .features, .shiftClickedFeatures, .labelledFeatures */
 
   /** Called when an axis feature track or feature triangle is clicked.
    * Toggle membership of the feature in one of the arrays : .features or .labelledFeatures
@@ -190,5 +213,73 @@ export default Service.extend(Evented, {
   }),
 
   /*--------------------------------------------------------------------------*/
+  /** Operations on .blocksFeatures */
+
+  updateSelectedFeatures: function(features) {
+    const fnName = 'updateSelectedFeatures';
+    // dLog(fnName, features.length);
+    this.set('selectedFeatures', features);
+  },
+
+  selectedFeatures_add(mapChrName, feature) {
+    this.blocksFeatures[mapChrName].push(feature);
+  },
+
+  selectedFeatures_clear()
+  {
+    const selectedFeatures = this.blocksFeatures;
+    /* delete properties instead of '= {}'
+     * to preserve the reference
+     */
+    for (const mapChrName in selectedFeatures){
+      if (selectedFeatures.hasOwnProperty(mapChrName)) {
+        delete selectedFeatures[mapChrName];
+      }
+    }
+    // no-op because selectedFeatures reference has not changed.
+    this.updateSelectedFeatures();
+  }
+
+
+  //----------------------------------------------------------------------------
 
 });
+
+//------------------------------------------------------------------------------
+
+function selectedBlocksFeaturesToArray(selectedFeatures) {
+  const
+  featuresAsArray = Object.keys(selectedFeatures)
+    .map(function (key) {
+      return selectedFeatures[key].map(function(feature) {
+        /** feature is now the Ember object models/feature 
+         * Until 0eeda0a7, feature contained feature name and position, separated by " ".
+         */
+        let selectedFeature = {
+          Chromosome : key,
+          Feature : feature.name,
+          Position : feature.location, /* i.e. .value[0]*/
+          /** Chromosome, Feature and Position can be derived from
+           * feature, so after the various uses of this are
+           * changed to use .feature, the structure can be
+           * replaced by simply feature.
+           */
+          feature
+        };
+        return selectedFeature;
+      });
+    })
+    .reduce(function(a, b) { 
+      return a.concat(b);
+    }, []);
+
+  return featuresAsArray;
+}
+
+//------------------------------------------------------------------------------
+
+export {
+  selectedBlocksFeaturesToArray,
+};
+
+//------------------------------------------------------------------------------
